# control_plane/api_spec.yaml
openapi: 3.1.0
info:
  title: Mission Control Remote Access — Control Plane API
  version: 0.1.0-draft
  description: |
    Authoritative HTTP API for `api.PLATFORM_DOMAIN` (the control plane in the
    MC remote-access architecture).

    Companion narrative: `docs/remote-access/03-control-plane-api.md`.
    Wire-format details:    `docs/remote-access/02-attestation-protocol.md`.
    System architecture:    `docs/remote-access/01-architecture.md`.

    The control plane never sees, proxies, or stores user dashboard contents.
    It is an identity, registry, and Cloudflare orchestration service.
  contact:
    name: Mission Control Remote Access
    email: support@PLATFORM_DOMAIN
  license:
    name: Proprietary
servers:
  - url: https://api.PLATFORM_DOMAIN/v1
    description: Production
  - url: https://api-staging.PLATFORM_DOMAIN/v1
    description: Staging
...
components:
  securitySchemes:
    FirebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token. Audience pinned to project id.

    DeviceSignature:
      type: apiKey
      in: header
      name: Authorization
      description: |
        `MC-Device device_id="<id>",sig_alg="ed25519"` — body carries a
        signed envelope (see protocol doc §7). The server fetches the device
        public key by `device_id`, verifies the wrapper signature, then
        validates the inner envelope against the 14-step checklist.

    OperatorJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Operator JWT issued via Google IAP. Audience-bound.
...
  responses:
...
    Unauthorized:
      description: Auth missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          description: Stable error code (see protocol doc §10 + API doc §7).
          examples: [bad_signature, rate_limited, version_floor_exceeded]
        message: { type: string, description: Human-readable, English. Clients localize via code. }
        request_id: { type: string }
        retry_after_ms: { type: integer, minimum: 0 }
        details: { type: object, additionalProperties: true }
...
    NonceResponse:
      type: object
      required: [nonce, nonce_id, expires_at]
      additionalProperties: false
      properties:
        nonce:
          type: string
          description: Base64 32 random bytes. Single-use.
        nonce_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
    AttestationEnvelope:
      type: object
      description: |
        Inner envelope. Canonical-JSON of this object is what gets signed
        (twice — by device key and by platform client secret).
        See protocol doc §7.1.
      required:
        - proto
        - envelope_type
        - device_pub_b64
        - device_id
        - enrollment_token
        - mc_version
        - mc_tunnel_version
        - client_secret_key_id
        - os
        - hostname_claim
        - timestamp
        - nonce
        - challenge_response
      additionalProperties: false
      properties:
        proto: { type: integer, enum: [1] }
        envelope_type: { type: string, enum: [attestation_request] }
        device_pub_b64: { type: string }
        device_id: { type: string }
        enrollment_token: { type: string }
        mc_version: { type: string }
        mc_tunnel_version: { type: string }
        client_secret_key_id:
          type: string
          description: |
            Identifies which platform client-secret keypair signed
            `client_signature_b64` in the wrapper. Must be in the active set
            (latest 3 non-revoked entries in `client_secret_keys` collection).
          examples: [mc-tunnel-2026a]
        os: { type: string }
        hostname_claim: { type: string }
        timestamp: { type: string, format: date-time }
        nonce: { type: string }
        challenge_response: { type: string }
        previous_token_id:
          type: string
          nullable: true
        # Reserved (re-enabled when code-signing returns to scope per 05- §1):
        build_manifest_id:
          type: string
          nullable: true
        mc_binary_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
          nullable: true
    AttestationRequest:
      type: object
      description: |
        Outer wrapper sent on the wire. Server recomputes
        sha256(canonical-JSON(envelope)) and verifies BOTH signatures:
        `signature_b64` (device key, proves which user/device) and
        `client_signature_b64` (platform client secret, proves it's a real
        platform-issued mc-tunnel — see protocol doc §3.6).
      required:
        - envelope
        - envelope_canonical_sha256
        - signature_b64
        - client_signature_b64
      additionalProperties: false
      properties:
        envelope: { $ref: '#/components/schemas/AttestationEnvelope' }
        envelope_canonical_sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        signature_b64:
          type: string
          description: Ed25519 signature over `envelope_canonical_sha256` by `device_priv`.
        client_signature_b64:
          type: string
          description: Ed25519 signature over `envelope_canonical_sha256` by `CLIENT_SECRET_PRIV` (baked into mc-tunnel).
    AttestationDirective:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [force_logout, update_required, notify_user, pause, step_up_auth]
        text: { type: string }
        seconds: { type: integer }
        reason: { type: string }
    AttestationCaps:
      type: object
      required:
        - bandwidth_bytes_remaining_period
        - rate_limit_rps
        - max_response_bytes
        - max_concurrent_connections
      properties:
        bandwidth_bytes_remaining_period: { type: integer }
        rate_limit_rps: { type: integer }
        max_response_bytes: { type: integer }
        max_concurrent_connections: { type: integer }
    AttestationResponse:
      type: object
      required:
        - envelope_type
        - result
        - tunnel_token
        - tunnel_token_id
        - tunnel_token_expires_at
        - next_attestation_after
        - caps
        - directives
      additionalProperties: false
      properties:
        envelope_type: { type: string, enum: [attestation_response] }
        result: { type: string, enum: [ok] }
        tunnel_token: { type: string }
        tunnel_token_id: { type: string, format: uuid }
        tunnel_token_expires_at: { type: string, format: date-time }
        next_attestation_after: { type: string, format: date-time }
        caps: { $ref: '#/components/schemas/AttestationCaps' }
        directives:
          type: array
          items: { $ref: '#/components/schemas/AttestationDirective' }

    #─── Admin ─────────────────────────────────────────────────────────────────
...
paths:
...
  /nonce:
    get:
      tags: [attestation]
      summary: Issue single-use attestation nonce
      security:
        - DeviceSignature: []
      parameters:
        - { name: device_id, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NonceResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /attest:
    post:
      tags: [attestation]
      summary: Verify envelope, issue tunnel token (or directive)
      description: |
        The 14-step verification list is in `02-attestation-protocol.md` §7.4.
        On success, returns a 15-minute Cloudflare tunnel token. On any failure,
        returns the protocol error code in the `code` field of the error envelope.

        Tunnel token rotation: client should re-attest every 10 minutes
        (5-minute safety margin before the 15-minute TTL) and pass the previous
        `tunnel_token_id` in `previous_token_id`.
      security:
        - DeviceSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttestationRequest' }
      responses:
        '200':
          description: Attestation accepted; tunnel token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttestationResponse' }
        '400':
          description: Malformed envelope, replay, hash mismatch, etc.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          description: bad_signature / unknown_device / bad_enrollment_token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '403':
          description: revoked_device / device_cap_exceeded / version_floor_exceeded / account_suspended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '410':
          description: revoked_build / unknown_build
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: Internal — retry with jitter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  #─── Webhooks ────────────────────────────────────────────────────────────────
...